requirements.txt
certifi==2023.7.22
chardet==3.0.4
GroupyAPI==0.10.3
idna==2.7
requests==2.20.1
urllib3==1.24.3
blinker==1.6.2
certifi==2023.7.22
chardet==3.0.4
click==8.1.7
Flask==2.3.3
GroupyAPI==0.10.3
idna==2.7
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.20.1
urllib3==1.24.3
Werkzeug==2.3.7
zipp==3.17.0
